modify database for different users

+++ MsgServerClientSdk/SdkMsgClientAndroid/sdkmsgclient/src/main/java/org/dync/teameeting/sdkmsgclient/msgs/MSSqlite.java
@@ -26,9 +26,9 @@ public class MSSqlite {
         return true;
     }
 
-    public boolean IsUserExistsInDb(String userId) {
+    public boolean IsUserExistsInDb(String userId, String seqnId) {
         SQLiteDatabase db = mSql3Db.getReadableDatabase();
-        String sqlSelect = "select count(*) from " + SqliteDbHelper.MS_TABLE_STOREID_SEQN + " where seqnId='" + userId +"';" ;
+        String sqlSelect = "select count(*) from " + SqliteDbHelper.MS_TABLE_STOREID_SEQN + " where userId='" + userId + "' and seqnId='" + seqnId +"';" ;
         Cursor cursor = db.rawQuery(sqlSelect, null);
         int code = 0;
         if (cursor.moveToFirst()) {
@@ -42,9 +42,10 @@ public class MSSqlite {
         return exist;
     }
 
-    public boolean InsertSeqn(String seqnId, long seqn) {
+    public boolean InsertSeqn(String userId, String seqnId, long seqn) {
         SQLiteDatabase db = mSql3Db.getWritableDatabase();
         ContentValues value = new ContentValues();
+        value.put("userId", userId);
         value.put("seqnId", seqnId);
         value.put("seqn", seqn);
         value.put("isfetched", 0);
@@ -53,28 +54,28 @@ public class MSSqlite {
         return true;
     }
 
-    public boolean UpdateSeqn(String seqnId, long seqn) {
+    public boolean UpdateSeqn(String userId, String seqnId, long seqn) {
         SQLiteDatabase db = mSql3Db.getWritableDatabase();
         ContentValues value = new ContentValues();
         value.put("seqn", seqn);
-        db.update(SqliteDbHelper.MS_TABLE_STOREID_SEQN, value, "seqnId=?", new String[]{seqnId});
+        db.update(SqliteDbHelper.MS_TABLE_STOREID_SEQN, value, "userId=? and seqnId=?", new String[]{userId, seqnId});
         db.close();
         return true;
     }
 
-    public boolean UpdateSeqn(String seqnId, long seqn, int isFetched) {
+    public boolean UpdateSeqn(String userId, String seqnId, long seqn, int isFetched) {
         SQLiteDatabase db = mSql3Db.getWritableDatabase();
         ContentValues value = new ContentValues();
         value.put("seqn", seqn);
         value.put("isfetched", isFetched);
-        db.update(SqliteDbHelper.MS_TABLE_STOREID_SEQN, value, "seqnId=?", new String[]{seqnId});
+        db.update(SqliteDbHelper.MS_TABLE_STOREID_SEQN, value, "userId=? and seqnId=?", new String[]{userId, seqnId});
         db.close();
         return true;
     }
 
-    public long SelectSeqn(String seqnId) {
+    public long SelectSeqn(String userId, String seqnId) {
         SQLiteDatabase db = mSql3Db.getReadableDatabase();
-        String sqlSelect = "select seqn from " + SqliteDbHelper.MS_TABLE_STOREID_SEQN + " where seqnId='" + seqnId +"';" ;
+        String sqlSelect = "select seqn from " + SqliteDbHelper.MS_TABLE_STOREID_SEQN + " where userId='" + userId + "' and seqnId='" + seqnId +"';" ;
         Cursor cursor = db.rawQuery(sqlSelect, null);
         long code = -1;
         if (cursor.moveToFirst()) {
@@ -87,25 +88,26 @@ public class MSSqlite {
         return code;
     }
 
-    public boolean DeleteSeqn(String seqnId) {
+    public boolean DeleteSeqn(String userId, String seqnId) {
         SQLiteDatabase db = mSql3Db.getWritableDatabase();
-        db.delete(SqliteDbHelper.MS_TABLE_STOREID_SEQN, "seqnId=?", new String[]{seqnId});
+        db.delete(SqliteDbHelper.MS_TABLE_STOREID_SEQN, "userId=? and seqnId=?", new String[]{userId, seqnId});
         db.close();
         return true;
     }
 
-    public boolean InsertGroupId(String grpId) {
+    public boolean InsertGroupId(String userId, String grpId) {
         SQLiteDatabase db = mSql3Db.getWritableDatabase();
         ContentValues value = new ContentValues();
+        value.put("userId", userId);
         value.put("seqnId", grpId);
         db.insert(SqliteDbHelper.MS_TABLE_GROUPS_ID, null, value);
         db.close();
         return true;
     }
 
-    public boolean IsGroupExistsInDb(String groupId) {
+    public boolean IsGroupExistsInDb(String userId, String groupId) {
         SQLiteDatabase db = mSql3Db.getReadableDatabase();
-        String sqlSelect = "select count(*) from " + SqliteDbHelper.MS_TABLE_GROUPS_ID + " where seqnId='" + groupId +"';" ;
+        String sqlSelect = "select count(*) from " + SqliteDbHelper.MS_TABLE_GROUPS_ID + " where userId='" + userId + "' and seqnId='" + groupId +"';" ;
         Cursor cursor = db.rawQuery(sqlSelect, null);
         int code = 0;
         if (cursor.moveToFirst()) {
@@ -119,18 +121,19 @@ public class MSSqlite {
         return exist;
     }
 
-    public ArrayList<GroupInfo> SelectGroupInfo() {
+    public ArrayList<GroupInfo> SelectGroupInfo(String userId) {
         SQLiteDatabase db = mSql3Db.getReadableDatabase();
-        String sqlSelect = "select * from " + SqliteDbHelper.MS_TABLE_STOREID_SEQN +";" ;
+        String sqlSelect = "select * from " + SqliteDbHelper.MS_TABLE_STOREID_SEQN +" where userId='" + userId + "';" ;
         Cursor cursor = db.rawQuery(sqlSelect, null);
         ArrayList<GroupInfo> arrInfo = new ArrayList<GroupInfo> ();
         if (cursor.moveToFirst()) {
             do {
                 HashMap<String, String> in = new HashMap<String, String>();
                 GroupInfo gi = new GroupInfo();
-                gi.setmSeqnId(cursor.getString(0));
-                gi.setmSeqn(cursor.getLong(1));
-                gi.setmIsFetched(cursor.getInt(2));
+                gi.setmUserId(cursor.getString(0));
+                gi.setmSeqnId(cursor.getString(1));
+                gi.setmSeqn(cursor.getLong(2));
+                gi.setmIsFetched(cursor.getInt(3));
                 arrInfo.add(gi);
             } while(cursor.moveToNext());
         }
@@ -139,9 +142,9 @@ public class MSSqlite {
         return arrInfo;
     }
 
-    public boolean DeleteGroupId(String grpId) {
+    public boolean DeleteGroupId(String userId, String grpId) {
         SQLiteDatabase db = mSql3Db.getWritableDatabase();
-        db.delete(SqliteDbHelper.MS_TABLE_GROUPS_ID, "seqnId=?", new String[]{grpId});
+        db.delete(SqliteDbHelper.MS_TABLE_GROUPS_ID, "userId=? and seqnId=?", new String[]{userId, grpId});
         db.close();
         return true;
     }
